AG-12556 - Simplify homepage gallery examples
