+++ packages/ag-charts-website/src/features/homepage/components/HomepageGalleryExamples.astro
@@ -1,0 +1,14 @@
+---
+import ExampleRunnerContainer from '@features/example-runner/components/ExampleRunnerContainer.astro';
+import { HomepageGalleryExamples } from './HomepageGalleryExamples.jsx';
+import { getLoadingIFrameId } from '@features/example-runner/utils/getLoadingLogoId';
+
+const { examples, pageName } = Astro.props;
+
+const exampleName = examples[0].exampleName;
+const loadingIFrameId = getLoadingIFrameId({ pageName, exampleName });
+---
+
+<ExampleRunnerContainer pageName={pageName} exampleName={exampleName}>
+    <HomepageGalleryExamples client:load examples={examples} loadingIFrameId={loadingIFrameId} />
+</ExampleRunnerContainer>
